fix(binary): Pull in length-prefix parsers
