+++ tests/testsuite/overflow.rs
@@ -3,10 +3,11 @@
 
 #[cfg(feature = "alloc")]
 use winnow::binary::be_u64;
+use winnow::binary::length_data;
 use winnow::bytes::take;
 use winnow::error::{ErrMode, Needed};
 #[cfg(feature = "alloc")]
-use winnow::multi::{length_data, many0};
+use winnow::multi::many0;
 use winnow::prelude::*;
 use winnow::Partial;
 
@@ -134,7 +135,7 @@ fn overflow_incomplete_count() {
 #[cfg(feature = "alloc")]
 fn overflow_incomplete_length_count() {
     use winnow::binary::be_u8;
-    use winnow::multi::length_count;
+    use winnow::binary::length_count;
 
     fn multi(i: Partial<&[u8]>) -> IResult<Partial<&[u8]>, Vec<&[u8]>> {
         length_count(be_u8, length_data(be_u64)).parse_next(i)
